+++ app/src/main/kotlin/org/btcmap/auth/RequestBuilderExt.kt
@@ -2,8 +2,8 @@ package org.btcmap.auth
 
 import okhttp3.Request
 
-class PublicRequest
+object PublicRequest
 
 fun Request.Builder.withoutAuth(): Request.Builder = apply {
-    tag(PublicRequest::class.java, PublicRequest())
+    tag(PublicRequest::class.java, PublicRequest)
 }